multitag-code/requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.9.0
torchvision==0.10.0
torchaudio==0.9.0
numpy=1.20.3
pandas==1.2.4
Pillow==7.0.0
opencv-python=4.5.2.54
tqdm=4.61.1
matplotlib=3.4.2
 
